remove backend module, move backend into mwbase, create an example mwhiv module with HIV overrides and swapped automatedmessage class
