+++ mwbase/models/__init__.py
@@ -5,4 +5,5 @@ from mwbase.models.misc import Connection, Practitioner, EventLog
 from mwbase.models.visit import Visit, ScheduledPhoneCall
 
 # Must be last since participants imports the others
+from mwbase.models.automatedmessage import AutomatedMessage, AutomatedMessageQuerySetBase, AutomatedMessageBase
 from mwbase.models.participants import Participant, StatusChange
