+++ mwbase/templates/admin/mwbase/automatedmessage/change_list.html
