+++ mwbase/templates/admin/mwbase/automatedmessage/sms_bank_import.html
@@ -27,7 +27,7 @@
 <div class="breadcrumbs">
 <a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
 &rsaquo; <a href="{% url 'admin:app_list' app_label=opts.app_label %}">{{ opts.app_config.verbose_name }}</a>
-&rsaquo; <a href="{% url 'admin:mwbase_message_changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
+&rsaquo; <a href="{% url 'admin:mwbase_automatedmessage_changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
 &rsaquo; SMS Bank Import
 </div>
 {% endblock %}
