+++ mwhiv/templates/admin/mwhiv/automatedmessagehiv/sms_bank_check.html
@@ -1,0 +1,11 @@
+{% extends 'admin/mwbase/automatedmessage/sms_bank_check.html' %}
+{% load i18n admin_urls static admin_list crispy_forms_tags %}
+
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+<a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
+&rsaquo; <a href="{% url 'admin:app_list' app_label=opts.app_label %}">{{ opts.app_config.verbose_name }}</a>
+&rsaquo; <a href="{% url 'admin:mwhiv_automatedmessagehiv_changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
+&rsaquo; SMS Bank Import Check
+</div>
+{% endblock %}
